int to long

+++ src/android/AudioBuffer.java
@@ -110,7 +110,7 @@ public class AudioBuffer extends CordovaPlugin {
                     //start calling run in a timertask
                     TimerTask timerTask = new TimerTask() {
                         public void run() {
-                            int s = System.currentTimeMillis();
+                            long s = System.currentTimeMillis();
                             int readSize = that.audioRecord.read(that.buffer, 0, that.buffer.length);
                             String toSend = "{ \"data\": " + Arrays.toString(that.buffer) + " }";
                             PluginResult result = new PluginResult(PluginResult.Status.OK, toSend );
